requirements.txt
Pillow==9.3.0
Pillow==9.5.0
streamlit==1.22.0
